+++ tests/test.py
@@ -9,6 +9,10 @@ import unittest
 from optimacore.project import Project
 from optimacore.framework_io import createFrameworkTemplate
 
+from optimacore.workbook import writeWorkbook
+from optimacore.system import SystemSettings as SS
+from optimacore.framework_settings import WorkbookSettings as FS
+
 import sys
 import pprint
 
@@ -32,10 +36,12 @@ class EverythingTest(unittest.TestCase):
 
     def test_template_process(self):
         """  """
-        createFrameworkTemplate(framework_path = self.framework_template_filepath)
-        self.proj.getFramework().importFromFile(framework_path = self.framework_template_filepath)
-        self.proj.createDatabook(databook_path = self.databook_template_filepath)
-        #pprint.pprint(self.proj.getFramework().specs)
+        writeWorkbook(workbook_path = "./frameworks/framework_test.xlsx", framework = None, data = None, instructions = None, workbook_type = SS.WORKBOOK_KEY_FRAMEWORK)
+        #createFrameworkTemplate(framework_path = self.framework_template_filepath)
+        #self.proj.getFramework().importFromFile(framework_path = self.framework_template_filepath)
+        #self.proj.createDatabook(databook_path = self.databook_template_filepath)
+        pprint.pprint(FS.PAGE_SPECS)
+        pprint.pprint(FS.ITEM_TYPE_SPECS)
         return None
     
     #def test_framework_example_import(self):