+++ CHANGELOG.md
@@ -3,10 +3,19 @@
 All notable changes to this project are documented in this file.
 The format adheres to Semantic Versioning.
 
+## 0.3.4 (2018-02-05): Workbook Construction Consolidation
+
+Features
+- Templated out a `ProjectData` structure to store model values for a corresponding `ProjectFramework` structure.
+
+Edits
+- Major change applied, with `FrameworkSettings` and `DatabookSettings` now deriving from new class `BaseStructuralSettings`.
+
 ## 0.3.3 (2018-02-01): Basic Databook Items
 
 Features
 - Improved wrapped `createDatabookFunc()` to produce a list of populations in a databook.
+- Characteristics and parameters are also listed out.
 
 Edits
 - Pulled Excel-based functionality and format variables into file: excel.py